Fixing undefined variable error
If there is more than one item in $categories, this would lead to an
undefined variable error for $categoryUserArray, since this variable
is unset at the end of each loop. A simple solution is to set it to
an empty array at the beginning of the loop. This fixes the notice
and does not affect the code in any other way. $categoryUserArray is
set to an empty array in line 753.